feat: split commands declarations
